% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib2.R
\name{calib2}
\alias{calib2}
\title{Calibration errors}
\usage{
calib2(
  Xs_old,
  d,
  total_old,
  q = rep(1, length(d)),
  method = c("linear", "raking", "truncated", "logit"),
  bounds = c(low = 0, upp = 10),
  description = FALSE,
  max_iter = 500,
  EPS1 = 1e-06,
  cut_off = 1,
  constant = TRUE,
  verbose = FALSE,
  relative_error = FALSE,
  digit_level = 6,
  print_indvar = FALSE
)
}
\arguments{
\item{Xs_old}{matrix of calibration variables.}

\item{d}{vector of initial weights.}

\item{total_old}{vector of population totals.}

\item{q}{vector of positive values accounting for heteroscedasticity; the variation of the g-weights is reduced for small values of q.}

\item{method}{calibration method (linear, raking, logit, truncated).}

\item{bounds}{vector of bounds for the g-weights used in the truncated and logit methods; 'low' is the smallest value and 'upp' is the largest value.}

\item{description}{if description=TRUE, summary of initial and final weights are printed, and their boxplots and histograms are drawn; by default, its value is FALSE.}

\item{max_iter}{maximum number of iterations in the Newton's method.}

\item{EPS1}{tolerance in the algorithm.(default is eps1=1e-06)}

\item{cut_off}{A numeric value for the pair-wise absolute correlation cutoff}

\item{constant}{if constant=TRUE, its remove constant variables in the matrix of calibration.}

\item{verbose}{A boolean for printing the details of function findCorrelations.}

\item{relative_error}{if relative_error=TRUE, checks quality of the relative error. If relative_error = FALSE, checks quality of the absolute error.}

\item{digit_level}{integer which shows how many digits rounds the results in the result table.}

\item{print_indvar}{if print_indvar=TRUE, Independent variables are printed; by default, its value is FALSE.}
}
\value{
Calculate maximum absolute error and maximum relative error.
}
\description{
Computes the g-weights of the calibration estimator.
}
\examples{
\donttest{
library(sampling)
Xs <- cbind(c(1,1,1,1,1,0,0,0,0,0),c(0,0,0,0,0,1,1,1,1,1),c(1,2,3,4,5,6,7,8,9,10))

}

}
\seealso{
\code{\link{Calibrationerrors}}
}
