% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibrationerrors.R
\name{Calibrationerrors}
\alias{Calibrationerrors}
\title{Calibration errors}
\usage{
Calibrationerrors(Xs, d, total, g)
}
\arguments{
\item{Xs}{matrix of calibration variables.}

\item{d}{vector of initial weights.}

\item{total}{vector of population totals.}

\item{g}{vector of g-weights.}
}
\value{
Calculate maximum absolute error and maximum relative error.
}
\description{
Calculate maximum absolute error and maximum relative error
}
\examples{
\donttest{
library(sampling)
Xs <- cbind(c(1,1,1,1,1,0,0,0,0,0),c(0,0,0,0,0,1,1,1,1,1),c(1,2,3,4,5,6,7,8,9,10))

}

}
\seealso{
\code{\link{calib2}}
}
