% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outlier.R
\name{detect_outlier}
\alias{detect_outlier}
\title{Calculated Outliers for single stage surveys}
\usage{
detect_outlier(
  id,
  vars,
  group,
  dataset,
  b = 1,
  a = 0.1,
  nd_log = 2,
  nd_log2 = 4,
  percthreshold = 0.95
)
}
\arguments{
\item{id}{(\code{character}) - unit id variable name.}

\item{vars}{(\code{character}) - variable name.}

\item{group}{(\code{character}) - group variable name.}

\item{dataset}{data.table;}

\item{b}{(\code{numeric}) - the user-dependent attitude to distances between succeeding values.}

\item{a}{(\code{numeric}) the threshold}

\item{nd_log}{(\code{numeric}) - correction for calculating average density of consecutive values}

\item{nd_log2}{(\code{numeric}) - second correction for calculating average density of consecutive values}

\item{percthreshold}{(\code{numeric}) - percentage of the threshold;}
}
\value{
dataset - survey data with outliers.
}
\description{
Detected outliers for single stage surveys.
}
\examples{
\donttest{
library(data.table)
}

}
\references{
Mark Last, Abraham Kandelm, Automated Detection of Outliers in Real-World Data, \url{http://www.ise.bgu.ac.il/faculty/mlast/papers/outliers2.pdf}
}
\seealso{
\code{\link{Calibrationerrors}}
}
