% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonresponse_correction.R
\name{nonresponse_correction}
\alias{nonresponse_correction}
\title{Nonresponse correction for single stage surveys}
\usage{
nonresponse_correction(
  in_system = "in_isdavs",
  sample = "izlase",
  frame = "frame",
  outlier = NULL,
  response_ind = "ievad",
  summary_w = "kops_sv",
  design_weight = "diz_sv",
  pop_size = NULL,
  new_pop_size = NULL,
  strata = "strata",
  dataset
)
}
\arguments{
\item{in_system}{(\code{character}) - indicator that tell us, is data in system.}

\item{sample}{(\code{character}) - sampling indicator.}

\item{frame}{(\code{character}) - frame indicator variable name.}

\item{outlier}{(\code{character}) - outlier indicator variable name.}

\item{response_ind}{(\code{character}) - response indicator variable name,  that indicates which respondents are weighted in the weight calculation process.}

\item{summary_w}{(\code{character}) summary weight indicator variable name, that indicates which respondents have a weight greater than zero. Weight will be used to calculate summaries.}

\item{design_weight}{(\code{character}) - design weight variable name.}

\item{pop_size}{(\code{character}) - population size in strata variable name.}

\item{new_pop_size}{(\code{character}) - updated frame population size in strata variable name.}

\item{strata}{(\code{character}) - strata variable name;}

\item{dataset}{data.table;}
}
\value{
dataset - survey data with weight (koef) and summary weight (koef_nodot).
}
\description{
Calcutate nonresponse correction factor for single stage surveys
}
\examples{
library(data.table)
\donttest{

}

}
\seealso{
\code{\link{Calibrationerrors}}
}
